extract todoConverter from page.tsx to a dedicated file
/lib/converters/todo-converter.ts
